Minor cleanups to reduce sprawl/improve CNI logging
